Moved integration tests to new directory
